ProyectosDjango/Django_Rest_API/requirements.txt
django==4.0
djangorestframework==3.13.0